[M6/M6TV] remove m6_skt_v1 m6_ref_v1 m6tv_skt_v1 m6tv_ref_v1 compile warnings

+++ arch/arm/cpu/aml_meson/m6tv/firmware/appf/os_api.c
@@ -107,8 +107,8 @@ int appf_boottime_init(void)
 		update_offset();
     appf_platform_boottime_init();
 
-    *((unsigned*)reloc_addr(&appf_runtime_call_flat_mapped)) =  appf_runtime_call;
-    *((unsigned*)reloc_addr(&appf_device_memory_flat_mapped)) = reloc_addr(appf_device_memory);
+    *((unsigned*)reloc_addr((unsigned)&appf_runtime_call_flat_mapped)) =  (unsigned)appf_runtime_call;
+    *((unsigned*)reloc_addr((unsigned)&appf_device_memory_flat_mapped)) = reloc_addr((unsigned)appf_device_memory);
 		appf_setup_translation_tables();
     return 0;
 }
@@ -135,7 +135,7 @@ int appf_runtime_init(void)
 {
     int ret,i;
     unsigned va, pa;
-    struct appf_main_table* pmaintable = (struct appf_main_table*)reloc_addr(&main_table);
+    struct appf_main_table* pmaintable = (struct appf_main_table*)reloc_addr((unsigned)&main_table);
     struct appf_cluster* pcluster;
     
     update_offset();
@@ -144,7 +144,7 @@ int appf_runtime_init(void)
     is_smp = FALSE;
 
     pa = reloc_addr((unsigned)&main_table);
-    va = &main_table;
+    va = (unsigned)&main_table;
     if (pa == va)
     {
     	 __V(flat_mapped) = TRUE;
@@ -230,7 +230,7 @@ static int power_down_cpu(unsigned cstate, unsigned rstate, unsigned flags)
     struct appf_cluster *cluster;
     int cpu_index, cluster_index;
     int i, rc, cluster_can_enter_cstate1;
-    struct appf_main_table* pmaintable = (struct appf_main_table*)reloc_addr(&main_table);
+    struct appf_main_table* pmaintable = (struct appf_main_table*)reloc_addr((unsigned)&main_table);
 #ifdef USE_REALVIEW_EB_RESETS
     int system_reset = FALSE, last_cpu = FALSE;
 #endif
@@ -240,7 +240,7 @@ static int power_down_cpu(unsigned cstate, unsigned rstate, unsigned flags)
     cluster = pmaintable->cluster_table;
     cluster += cluster_index;
     
-    dbg_print("cluster:",cluster);
+    dbg_print("cluster:",(int)cluster);
     
     cpu = cluster->cpu_table;
     cpu += cpu_index;
